backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-cosmos
azure-storage-blob
openai
